requirements.txt
django>=3.0
djangorestframework
mysqlclient
pillow
requests
django-cors-headers
uwsgi